+++ handler/endpoint.go
@@ -109,12 +109,21 @@ func (e *Endpoint) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 			if b.Err == nil {
 				continue
 			}
+
 			switch b.Err.(type) {
 			case producer.ResultPanic:
-				log.Error(b.Err)
+				log.WithError(b.Err).Error()
+			}
+
+			if b.Err != nil {
+				err = b.Err
+				break
 			}
 		}
-		clientres, err = producer.NewResponse(req, e.opts.Response.Context, evalContext, http.StatusOK)
+
+		if err == nil {
+			clientres, err = producer.NewResponse(req, e.opts.Response.Context, evalContext, http.StatusOK)
+		}
 	} else {
 		if result, ok := beresps["default"]; ok {
 			clientres = result.Beresp
@@ -149,7 +158,7 @@ func (e *Endpoint) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 			}
 		case producer.ResultPanic:
 			serveErr = errors.Server.With(err)
-			log.Error(err)
+			log.WithError(err).Error()
 		}
 		e.opts.Error.ServeError(serveErr).ServeHTTP(rw, req)
 		return
@@ -157,7 +166,7 @@ func (e *Endpoint) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 
 	// always apply before write: redirect, response
 	if err = eval.ApplyResponseContext(evalContext, e.opts.Context, clientres); err != nil {
-		log.Error(err)
+		log.WithError(err).Error()
 	}
 
 	select {